requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
boto3==1.34.52          # AWS S3 Integration
pymongo==4.6.1          # MongoDB Client
confluent-kafka==2.3.0  # Kafka Producer & Consumer
google-generativeai==0.5.0  # Gemini AI API
pydantic==2.6.4         # Data validation for FastAPI
python-dotenv==1.0.1    # Environment variables
kafka-python


#import os

# Load environment variables from .env file


python-dotenv
